Retune for gfx110x for small messages, and address comments

+++ projects/rccl/src/include/rccl_graph_gen.h
@@ -33,7 +33,7 @@ RCCL_PARAM_DECLARE(InterGraphGen);
 void permute_array_inplace(int* input, int length, int* permutation);
 void findRingCutIndices(int nChannels, int nNodes /*nodes in the ring graph*/,
                         const int* flattenedRings /* Hamiltonian rings*/, int* cutIndices);
-ncclResult_t generateRings(int nNodes, uint8_t nChannels, int* nodeOrder);
+ncclResult_t generateRings(int nNodes, uint32_t nChannels, int* nodeOrder);
 
 #endif  // RCCL_GRAPH_GEN_H_
